requirements.txt
python-dotenv
youtube-transcript-api
groq
langchain
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
